fixed error for 5-flip_bits.c

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -18,7 +18,7 @@
 
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	unsigned int tmp = n ^ m, count = 0;
+	unsigned long tmp = n ^ m, count = 0;
 
 	while (tmp)
 	{
